moves metric package on service level
