+++ service/kvmconfig/v4/resource/deployment/current.go
@@ -11,7 +11,7 @@ import (
 	"k8s.io/api/extensions/v1beta1"
 	apismetav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/giantswarm/kvm-operator/service/kvmconfig/metric"
+	"github.com/giantswarm/kvm-operator/service/metric"
 	"github.com/giantswarm/kvm-operator/service/kvmconfig/v4/key"
 )
